requirement.txt
absl-py==2.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
attrs==24.3.0
backoff==2.2.1
blinker==1.9.0
branca==0.8.1
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.6
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
flatbuffers==25.2.10
folium==0.19.4
fonttools==4.55.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
htmlmin==0.1.12
idna==3.10
ImageHash==4.3.1
iniconfig==2.0.0
jax==0.5.2
jaxlib==0.5.1
Jinja2==3.1.5
joblib==1.1.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.3
mediapipe==0.10.21
missingno==0.5.2
ml_dtypes==0.5.1
MouseInfo==0.1.3
multidict==6.1.0
multimethod==2.0
mysql-connector==2.2.9
narwhals==1.32.0
networkx==3.4.2
numpy==1.26.4
opencage==3.0.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.1
pandas==2.2.3
pandas-profiling==3.2.0
phik==0.12.4
phonenumbers==9.0.0
pillow==11.0.0
pluggy==1.5.0
propcache==0.3.0
protobuf==4.25.6
pyarrow==19.0.1
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==1.10.2
pydantic_core==2.27.2
pydeck==0.9.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.3.3
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.2
sounddevice==0.5.1
streamlit==1.44.0
tangled-up-in-unicode==0.2.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
visions==0.7.4
watchdog==6.0.0
xyzservices==2024.9.0
yarl==1.18.3
